springMVCFormLabel

+++ SpringMVCFormLabel/src/main/java/com/service/BookService.java
@@ -11,6 +11,6 @@ public interface BookService {
     List<Book> getAllBooks();//获取全部的书
     Book save(Book book);//保存书
     Book update(Book book);  //更新书
-    Book get(int id);//通过id来找书
+    Book get(long id);//通过id来找书
     long getNextId();//获取下一本书的id，用来添加书时自动给书赋id
 }
